# contracts/openapi/platform-v1.yaml
openapi: 3.1.0
info:
  title: Wealth Sandbox Platform API
  version: 1.0.0
  description: Canonical contract consumed by the mobile application.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Current platform API origin
paths:
...
  /api/v1/orders/preview:
    post:
      operationId: previewBuyOrder
      summary: Create an immutable synthetic BUY quote
      description: Requires `order.execute`; non-owned account/instrument combinations return 404.
      security:
        - BearerAuth: []
      tags: [Trading]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuoteRequest'
      responses:
        '201':
          description: Synthetic quote valid for 60 seconds.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/orders:
    get:
      operationId: listOrders
      summary: List current user's BUY orders
      description: Requires `financial.read`; cursor is the last order ID from the prior page.
      security:
        - BearerAuth: []
      tags: [Trading]
      parameters:
        - name: cursor
          in: query
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Owner-scoped order page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: prepareBuyOrder
      summary: Idempotently create a BUY order and reserve cash
      description: Requires `order.execute`. Cash reservation commits before one non-retried external POST.
      security:
        - BearerAuth: []
      tags: [Trading]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: Existing result for the same key and canonical payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '201':
          description: New order reached FILLED or REJECTED deterministically.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '202':
          description: External result is UNKNOWN and reconciliation is queued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Idempotency conflict, expired quote, or insufficient funds.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/orders/{orderId}:
    get:
      operationId: getOrder
      summary: Get one owner-scoped order
      description: Requires `financial.read`.
      security:
        - BearerAuth: []
      tags: [Trading]
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Current persisted order state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Bearer token is missing, expired, or cannot be verified.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Bearer token does not include the required scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ValidationFailed:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource does not exist or is not owned by the caller.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    Money:
      type: string
      pattern: '^-?[0-9]+\.[0-9]{4}$'
...
    CreateQuoteRequest:
      type: object
      additionalProperties: false
      required: [accountId, instrumentId, side, quantity]
      properties:
        accountId:
          type: string
          format: uuid
        instrumentId:
          type: string
          format: uuid
        side:
          type: string
          const: BUY
        quantity:
          type: string
          pattern: '^(?:0|[1-9][0-9]*)(?:\.[0-9]{1,8})?$'
    Quote:
      type: object
      additionalProperties: false
      required:
        - quoteId
        - side
        - quantity
        - unitPrice
        - estimatedAmount
        - fee
        - currency
        - expiresAt
        - syntheticQuote
      properties:
        quoteId:
          type: string
          format: uuid
        side:
          type: string
          const: BUY
        quantity:
          type: string
          pattern: '^[0-9]+\.[0-9]{8}$'
        unitPrice:
          $ref: '#/components/schemas/Money'
        estimatedAmount:
          $ref: '#/components/schemas/Money'
        fee:
          $ref: '#/components/schemas/Money'
        currency:
          type: string
          const: KRW
        expiresAt:
          type: string
          format: date-time
        syntheticQuote:
          type: boolean
          const: true
    CreateOrderRequest:
      type: object
      additionalProperties: false
      required: [quoteId, accountId, instrumentId, side, quantity]
      properties:
        quoteId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        instrumentId:
          type: string
          format: uuid
        side:
          type: string
          const: BUY
        quantity:
          type: string
          pattern: '^(?:0|[1-9][0-9]*)(?:\.[0-9]{1,8})?$'
    Order:
      type: object
      additionalProperties: false
      required:
        - orderId
        - status
        - side
        - quantity
        - estimatedAmount
        - filledAmount
        - createdAt
        - updatedAt
        - statusRefreshRecommendedAfterMs
      properties:
        orderId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING_SUBMISSION, UNKNOWN, FILLED, REJECTED, FAILED]
        side:
          type: string
          const: BUY
        quantity:
          type: string
          pattern: '^[0-9]+\.[0-9]{8}$'
        estimatedAmount:
          $ref: '#/components/schemas/Money'
        filledAmount:
          type: [string, 'null']
          pattern: '^[0-9]+\.[0-9]{4}$'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        statusRefreshRecommendedAfterMs:
          type: [integer, 'null']
          enum: [2000, null]
    OrderPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        nextCursor:
          type: [string, 'null']
          format: uuid
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - code
        - detail
        - traceId
        - retryable
        - fieldErrors
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
        detail:
          type: string
        traceId:
          type: string
        retryable:
          type: boolean
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - field
              - code
              - message
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
...